pyproject.toml
[tool.poetry]
name = "kinopoisk-bot"
version = "0.1.0"
description = ""
authors = ["yan0f <31446439+yan0f@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "kinopoisk_bot" }]

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.6"
httpx = "^0.25.0"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.format]
quote-style = "single"
